LayoutTests/http/tests/inspector/network/xhr-response-body-expected.txt

Tests for getting the content of XHR requests.


== Running test suite: Network.getResponseBody.XHR
-- Running test case: Network.getResponseBody.XHR.Text
PASS: Resource should be XHR type.
PASS: MIMEType should be 'text/plain'.
PASS: Content should not be base64Encoded.
PASS: Text content should match data.txt.

-- Running test case: Network.getResponseBody.XHR.HTML
PASS: Resource should be XHR type.
PASS: MIMEType should be 'text/html'.
PASS: Content should not be base64Encoded.
PASS: Text content should match data.html.

-- Running test case: Network.getResponseBody.XHR.JSON
PASS: Resource should be XHR type.
PASS: MIMEType should be 'application/octet-stream'.
PASS: Content should be base64Encoded.
PASS: JSON content should match data.json.

-- Running test case: Network.getResponseBody.XHR.JSON2
PASS: Resource should be XHR type.
PASS: MIMEType should be 'application/json'.
PASS: Content should not be base64Encoded.
PASS: JSON content should match specified content.

-- Running test case: Network.getResponseBody.XHR.JSON3
PASS: Resource should be XHR type.
PASS: MIMEType should be 'application/vnd.api+json'.
PASS: Content should not be base64Encoded.
PASS: JSON content should match specified content.

-- Running test case: Network.getResponseBody.XHR.SVG
PASS: Resource should be XHR type.
PASS: MIMEType should be 'image/svg+xml'.
PASS: Content should not be base64Encoded.
PASS: SVG content should be text.

-- Running test case: Network.getResponseBody.XHR.PNG
PASS: Resource should be XHR type.
PASS: MIMEType should be 'image/png'.
PASS: Content should be base64Encoded.
PASS: Image content should be a Blob.

-- Running test case: Network.getResponseBody.XHR.Sync.Text
PASS: Resource should be XHR type.
PASS: MIMEType should be 'text/plain'.
PASS: Content should not be base64Encoded.
PASS: Text content should match data.txt.

-- Running test case: Network.getResponseBody.XHR.Sync.PNG
PASS: Resource should be XHR type.
PASS: MIMEType should be 'image/png'.
PASS: Content should be base64Encoded.
PASS: Image content should be a Blob.
